erro

+++ SistemaCinema/src/br/unitins/cinema/autentication/Autenticacao.java
@@ -3,6 +3,6 @@ package br.unitins.cinema.autentication;
 public interface Autenticacao {
 	
 	boolean verificaSenha(String senha);
-	boolean verificaLogin(String login, Long id);
+	boolean verificaLogin(String login, Integer id);
 	
 }
